refactor(tests): reorganize test utilities by creating dedicated utility files for course, enrollment, user, progress tracking, and verification; remove unused common file
