+++ backend/src/modules/users/tests/utils/verifyProgressInDatabase.ts
